ensure sorting on a dynamic column in a partitioned table works in all cases
